% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeFixedPanelHybrid.R
\name{adaptive.lasso}
\alias{adaptive.lasso}
\title{Adaptive Lasso Variable Selection}
\usage{
adaptive.lasso(y, x)
}
\arguments{
\item{y}{Response vector.}

\item{x}{Design matrix.}
}
\value{
A vector of adaptive lasso coefficients.
}
\description{
Performs adaptive lasso variable selection using ridge regression for initial weight estimation.
}
