% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakDiagnostic.R
\name{BreakDiagnostic}
\alias{BreakDiagnostic}
\title{Detect a break number using different metrics}
\usage{
BreakDiagnostic(
  y,
  X,
  mcmc = 100,
  burn = 100,
  verbose = 100,
  thin = 1,
  break.upper = 3
)
}
\arguments{
\item{y}{Response vector.}

\item{X}{Design matrix.}

\item{mcmc}{The number of MCMC iterations after burn-in. Default is 100.}

\item{burn}{The number of burn-in iterations. Default is 100.}

\item{verbose}{Print progress every \code{verbose}th iteration. Default is 100.}

\item{thin}{Thinning interval. Default is 1.}

\item{break.upper}{Upper limit for break number detection. Default is 3.}
}
\value{
A list containing a ggplot diagnostic graph and a list of model comparison metrics.
}
\description{
Fits models with 0 to \code{break.upper} breaks and compares them using
log marginal likelihood, log likelihood, WAIC, and average loss.
}
\examples{
   \dontrun{
   set.seed(1973)
   ## One break test
   out <- BridgeChangeSim(ntime=20, predictor = 10,
     n.break=1, constant.p=0, varying.p=0.4, dgp.only=TRUE)

   ## Fit multiple models for break number detection
   detect <- BreakDiagnostic(y=out$y.c, X=out$x.c)

   ## Look at the graph
   detect[[1]]; print(detect[[2]])

   ## Two break test
   out <- BridgeChangeSim(ntime=20, predictor = 10,
     n.break=2, constant.p=0, varying.p=0.4, dgp.only=TRUE)

   ## Fit multiple models for break number detection using Bayesian model comparison
   detect <- BreakDiagnostic(y=out$y.c, X=out$x.c)
   
   ## Look at the graph
   detect[[1]]; print(detect[[2]])
  
}


}
