% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakPointLoss.R
\name{BreakPointLoss}
\alias{BreakPointLoss}
\title{Compute the Average Loss of Hidden State Changes from Expected Break Points}
\usage{
BreakPointLoss(model.list, marginal = FALSE, display = TRUE)
}
\arguments{
\item{marginal}{If \code{marginal} is TRUE, \code{logmarglike} will be reported.}

\item{display}{If \code{display} is TRUE, a plot of \code{ave.loss} will be produced. 

\code{BreakPointLoss}. ave.loss, logmarglike, State, Tau, Tau.samp}

\item{...}{MCMC output objects. These have to be of class
\code{mcmc} and have a \code{logmarglike} attribute. In what
follows, we let \code{M} denote the total number of models to be
compared.}
}
\value{
\code{BreakPointLoss} returns five objects. They are: \code{ave.loss} the expected loss for each model
  computed by the mean sqaured distance of hidden state changes from the expected break points
\textrm{Average Loss} = \frac{1}{M}\sum_{m=1}^{M}\left(\frac{1}{G}\sum_{g=1}^{G} (\bar{\tau}_m - \tau_{m}^{(g)})^2 \right);
  \code{logmarglike} the natural log of the marginal likelihood for each model; \code{State} sampled state vectors;
  \code{Tau} expected break points for each model; and \code{Tau.samp} sampled break points from hidden state draws.
}
\description{
Compute the Average Loss of Hidden State Changes from Expected Break Points
}
\examples{
\dontrun{
set.seed(1119)
n <- 99
ns <- 3
x1 <- runif(n)
true.beta1 <- c(2, -2)
true.beta2 <- c(-2,  -2)
true.beta3 <- c(0,  2)
true.Sigma <- c(1, 2)
true.s <- rep(1:ns, each=n/ns)

mu1 <- cbind(1, x1[true.s==1])\%*\%true.beta1
mu2 <- cbind(1, x1[true.s==2])\%*\%true.beta2
mu3 <- cbind(1, x1[true.s==3])\%*\%true.beta3

y <- as.ts(c(rnorm(n/ns, mu1, sd=sqrt(true.Sigma[1])), 
            rnorm(n/ns, mu2, sd=sqrt(true.Sigma[2])), 
             rnorm(n/ns, mu3, sd=sqrt(true.Sigma[1]))))
formula=y ~ x1
## prior
b0 <- 0; B0 <- 1
sigma.mu=sd(y)
sigma.var=var(y)
mcmc = 1000
## models
model1 <-  MCMCregressChange(formula, m=1, b0=b0, B0=B0, mcmc=mcmc, burnin=mcmc,
          sigma.mu=sigma.mu, sigma.var=sigma.var, marginal.likelihood="Chib95")
model2 <-  MCMCregressChange(formula, m=2, b0=b0, B0=B0, mcmc=mcmc, burnin=mcmc,
           sigma.mu=sigma.mu, sigma.var=sigma.var, marginal.likelihood="Chib95")
model3 <-  MCMCregressChange(formula, m=3, b0=b0, B0=B0, mcmc=mcmc, burnin=mcmc,
           sigma.mu=sigma.mu, sigma.var=sigma.var, marginal.likelihood="Chib95")
model4 <-  MCMCregressChange(formula, m=4, b0=b0, B0=B0, mcmc=mcmc, burnin=mcmc,
           sigma.mu=sigma.mu, sigma.var=sigma.var, marginal.likelihood="Chib95")
model5 <-  MCMCregressChange(formula, m=5, b0=b0, B0=B0, mcmc=mcmc, burnin=mcmc,
           sigma.mu=sigma.mu, sigma.var=sigma.var, marginal.likelihood="Chib95")

out <- BreakPointLoss(model1, model2, model3, model4, model5, marginal=TRUE)

print(out[["ave.loss"]])
}


}
