% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChange.R
\docType{package}
\name{BridgeChange-package}
\alias{BridgeChange}
\alias{BridgeChange-package}
\title{Fit Bayesian Bridge model with Change-points.}
\description{
This package implements gibbs updates for Bayesian sparse linear mixed effect model 
 for panel data with change-ponts.
}
\details{
This package offers the following functions:
\itemize{
  \item \code{\link{BridgeChangeReg}} Univariate Regression with Change point.
}
}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{syamauchi@princeton.edu}

Authors:
\itemize{
  \item Jong Hee Park \email{jongheepark@snu.ac.kr}
}

}
