% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeSim.r
\name{BridgeChangeSim}
\alias{BridgeChangeSim}
\title{Bridge Regression simulation}
\usage{
BridgeChangeSim(
  ntime = 500,
  predictor = 100,
  rho = 0,
  time.series = FALSE,
  standardize = TRUE,
  sign.change.tune = 2,
  sigma1 = 1,
  sigma2 = 2,
  train.ratio = 0.5,
  fitted.mse = TRUE,
  constant.p = 0.1,
  varying.p = 0.2,
  break.point = 0.5,
  positive.jump = FALSE,
  n.break = 1,
  intercept = FALSE,
  positive.jump.tune = 1,
  mcmc = 100,
  burn = 100,
  verbose = 100,
  thin = 1,
  N = 1,
  known.alpha = FALSE,
  dgp.only = FALSE
)
}
\arguments{
\item{ntime}{Length of time series}

\item{predictor}{Number of predictor}

\item{rho}{correlation parameter 0 = no correlation.}

\item{time.series}{TRUE if dgp is generated from autocorrelated series. rho is used as an autocorrelation coefficient.}

\item{standardize}{If TRUE, data are standardized. Default is TRUE.}

\item{sign.change.tune}{tuning parameter for the size of parameter sign changes}

\item{sigma1}{sigma 1}

\item{sigma2}{sigma 2}

\item{train.ratio}{The proportion of training data. (0, 1).}

\item{fitted.mse}{If TRUE and n.break == 0, compute the MSE of fitted values against true responses. If FALSE, do the cross-validation test using training data.}

\item{constant.p}{Proportion of constant parameters.}

\item{varying.p}{Proportion of time-varying parameters.}

\item{break.point}{Timing of a break between 0 and 1.}

\item{positive.jump}{If TRUE, breaks have positive jumps only.}

\item{n.break}{Number of structural breaks. Default is 1.}

\item{intercept}{If TRUE, include intercept. Default is FALSE.}

\item{positive.jump.tune}{Tuning parameter for positive jump size.}

\item{mcmc}{Number of MCMC iterations. Default is 100.}

\item{burn}{Number of burn-in iterations. Default is 100.}

\item{verbose}{Print progress every verbose-th iteration. Default is 100.}

\item{thin}{Thinning interval.}

\item{N}{Number of cross-sectional units. If \code{N=1}, univariate time series data.}

\item{known.alpha}{If TRUE, alpha is fixed. Default is FALSE.}

\item{dgp.only}{If TRUE, only data are generated and estimation steps are skipped.}
}
\value{
A list containing generated data and optionally model fit results.
}
\description{
Simulation code for univariate response change-point model with Bridge prior.
}
