% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeFixedPanelHybrid.R
\name{BridgeFixedPanelHybrid}
\alias{BridgeFixedPanelHybrid}
\title{Hybrid Approach to Bridge Change Point Model with Fixed Effect}
\usage{
BridgeFixedPanelHybrid(
  formula,
  data,
  index,
  model,
  effect,
  standardize = TRUE,
  inter = FALSE,
  n.break = 1,
  mcmc = 100,
  burn = 100,
  verbose = 100,
  thin = 1,
  b0 = 0,
  B0 = 1,
  c0 = 0.1,
  d0 = 0.1,
  r0 = 1,
  R0 = 1,
  nu.shape = 2,
  nu.rate = 2,
  alpha = 1,
  Waic = FALSE,
  marginal = FALSE
)
}
\arguments{
\item{data}{Data.frame object.}

\item{index}{String vector for unit and time index variables.
For example, \code{index = c("unit", "year")}.}

\item{model}{Model (\code{c("within","between", "pooling")}).}

\item{effect}{Effect (\code{c("individual", "time", "twoways")}).}

\item{standardize}{If TRUE, all covariates are standardized.}

\item{n.break}{Number of breaks.
If \code{n.break = 0}, it simply runs fixed effect model with shrinkage prior on coefficients.}

\item{mcmc}{MCMC iteration.}

\item{burn}{Burn-in period.}

\item{verbose}{Verbose.}

\item{thin}{Thinning.}

\item{c0}{Hyperparam}

\item{d0}{= 0.1}

\item{nu.shape}{=2.0}

\item{nu.rate}{=2.0}

\item{alpha}{= 1}

\item{fomula}{Inherited from \code{lm}. For example, \code{Y ~ X + Z}.}

\item{interaction}{If TRUE, all pairwise interactions are added.}

\item{ols.weight}{If TRUE, OLS estimates are used for adpative lasso weight vector.}

\item{sparse.only}{If TRUE, skip MCMC and fit a sparse regressiong using adaptive lasso. This works only for n.break=0.}
}
\description{
Hybrid Approach to Bridge Change Point Model with Fixed Effect
}
\author{
Jong Hee Park, and Soichiro Yamauchi \email{syamauchi@princeton.edu}
}
