% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.BridgeChange}
\alias{coef.BridgeChange}
\title{Obtain posterior mean of coefficients}
\usage{
\method{coef}{BridgeChange}(object, ...)
}
\arguments{
\item{object}{\code{BridgeChange} objects. Typically output from \code{BridgeChangeReg}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Obtain posterior mean of coefficients
}
