% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotRegime.R
\name{dotplotRegime}
\alias{dotplotRegime}
\title{Draw a plot of covariate-specific time-varying movements}
\usage{
dotplotRegime(
  out,
  hybrid = TRUE,
  start = 1,
  cex = 1,
  smooth.beta = TRUE,
  SE = FALSE,
  x.location = c("none", "random", "legend", "default"),
  order.state = 1,
  location.bar = 9,
  distance = 4,
  h = c(255, 330),
  l = c(40, 90),
  text.cex = 1,
  legend.position = "topright",
  pos = 1,
  select = NULL,
  main = "",
  raw = FALSE
)
}
\arguments{
\item{out}{Output from a BridgeChange model.}

\item{hybrid}{If TRUE, DSS results are used for graph. Otherwise, HMBB results are used.}

\item{start}{Starting value of the time sequence.}

\item{cex}{Point size. Default is 1.}

\item{smooth.beta}{If TRUE, probabilistic estimates of hidden states (smooth) are used. Otherwise, discrete state values are used.}

\item{SE}{If TRUE, draw 95 percent confidence interval.}

\item{x.location}{Location of covariate name labels. One of \code{"none"}, \code{"random"}, \code{"legend"}, or \code{"default"}.}

\item{order.state}{Regime index used to order transparency of coefficients. Default is 1.}

\item{location.bar}{Bar location parameter. Default is 9.}

\item{distance}{Distance of text labels from the maximum value of x. Default is 4.}

\item{h}{Hue value in the HCL color description, in [0, 360]. Default is c(255, 330).}

\item{l}{Luminance value in the HCL color description. Default is c(40, 90).}

\item{text.cex}{Text size for labels. Default is 1.}

\item{legend.position}{Position of legend. Default is \code{"topright"}.}

\item{pos}{Text position relative to point (1=below, 3=above). Default is 1.}

\item{select}{Character string to select specific coefficients by name. Default is NULL.}

\item{main}{Plot title. Default is empty string.}

\item{raw}{If TRUE, use raw y for variable selection. Default is FALSE.}
}
\description{
Draw a plot of covariate-specific time-varying movements
}
