% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{estimate_intercept_reg}
\alias{estimate_intercept_reg}
\title{Function to update intercept}
\usage{
estimate_intercept_reg(y, Xorig, beta, n.break, intercept, state)
}
\arguments{
\item{y}{outcome vector}

\item{Xorig}{design matrix in original scale}

\item{beta}{regression coefficient}

\item{n.break}{number of breaks}

\item{intercept}{boolean; if \code{FALSE}, the intercept is set to zero.}

\item{state}{state vector (length T)}
}
\description{
Function to update intercept
}
