% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse_state_sampler_NB}
\alias{sparse_state_sampler_NB}
\title{Negative Binomial State sampler}
\usage{
sparse_state_sampler_NB(m, y, X, d, beta, P)
}
\arguments{
\item{m}{Number of break points.}

\item{y}{Response vector.}

\item{X}{Design matrix.}

\item{d}{Dispersion parameter.}

\item{beta}{Coefficient matrix.}

\item{P}{Transition probability matrix.}
}
\description{
Negative Binomial State sampler
}
\keyword{internal}
