% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{WaicCompare}
\alias{WaicCompare}
\title{Compare Models based on WAIC}
\usage{
WaicCompare(outlist, print = TRUE)
}
\arguments{
\item{outlist}{list of \code{BridgeChange} objects, typically outputs from \code{BridgeChangeReg}.}

\item{print}{boolean; if \code{TRUE} selected model is printed.}
}
\value{
A data.frame of WAIC values.
}
\description{
Compare Models based on WAIC
}
