% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeFixedPanelHybrid.R
\name{adaptive.lasso}
\alias{adaptive.lasso}
\title{Hybrid Approach to Bridge Change Point Model with Fixed Effect}
\usage{
adaptive.lasso(y1, x1, beta.hat)
}
\arguments{
\item{fomula}{Inherited from \code{lm}. For example, \code{Y ~ X + Z}.}

\item{data}{Data.frame object.}

\item{index}{String vector for unit and time index variables.
For example, \code{index = c("unit", "year")}.}

\item{model}{Model (\code{c("within","between", "pooling")}).}

\item{effect}{Effect (\code{c("individual", "time", "twoways")}).}

\item{n.break}{Number of breaks.
If \code{n.break = 0}, it simply runs fixed effect model with shrinkage prior on coefficients.}

\item{mcmc}{MCMC iteration.}

\item{burn}{Burn-in period.}

\item{verbose}{Verbose.}

\item{thin}{Thinning.}

\item{c0}{Hyperparam}

\item{d0}{= 0.1}

\item{nu.shape}{=2.0}

\item{nu.rate}{=2.0}

\item{alpha}{= 1}
}
\description{
Hybrid Approach to Bridge Change Point Model with Fixed Effect
}
\author{
Jong Hee Park, and Soichiro Yamauchi \email{syamauchi@princeton.edu}
}
