% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChange.R
\docType{package}
\name{BridgeChange-package}
\alias{BridgeChange}
\alias{BridgeChange-package}
\title{Fit Bayesian Bridge model with Change-points.}
\description{
This package provides Bayesian implementation of Bridge regression with possible change-points.
The package supports univariate time-series data as well as panel data with linear and non-linear link functions.
In addition to the Gaussian linear models, count and binary outcomes are supported.
}
\details{
Specifically, this package offers the following functions:
\itemize{
  \item \code{\link{BridgeChangeReg}}: Univariate Regression with Change-points.
  \item \code{\link{BridgeFixedPanel}}: Fixed Effect Panel Regression with Change-points.
}
}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{syamauchi@g.harvard.edu}

Authors:
\itemize{
  \item Jong Hee Park \email{jongheepark@snu.ac.kr}
}

}
