% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeRegHybrid.r
\name{BridgeChangeRegHybrid}
\alias{BridgeChangeRegHybrid}
\title{Hybrid Bridge Regression with Change-point}
\usage{
BridgeChangeRegHybrid(
  y,
  X,
  n.break = 0,
  scale.data = TRUE,
  intercept = TRUE,
  mcmc = 100,
  burn = 100,
  verbose = 100,
  thin = 1,
  reduce.mcmc = NULL,
  ols.weight = FALSE,
  c0 = 0.1,
  d0 = 0.1,
  nu.shape = 2,
  nu.rate = 2,
  known.alpha = FALSE,
  alpha.start = 1,
  alpha.limit = FALSE,
  alpha.MH = TRUE,
  beta.start = NULL,
  beta.alg = "SVD",
  regime.duration.min = 5,
  waic = FALSE,
  marginal = FALSE
)
}
\arguments{
\item{y}{Outcome vector.}

\item{X}{Design matrix. Columns correspond to variables.}

\item{n.break}{The number of change-point(s).
If \code{n.break = 0}, the model corresponds to the usual regression.
Default is \code{n.break = 0}.}

\item{scale.data}{If \code{TRUE}, \code{y} and \code{X} are both scaled to have zero mean and unit variance.
Default is \code{TRUE}.
We recommend \code{scale.data = TRUE} unless the original data are already scaled.}

\item{intercept}{If \code{TRUE}, estimate intercept by MLE.
Default is \code{TRUE}.
This option does not affect the result when \code{n.break = 0}.
We recommend \code{intercept = TRUE} when the number of break is not zero.}

\item{mcmc}{The number of iterations for gibbs updates. Default is 100.}

\item{burn}{The number of burn-in periods for gibbs updates. Default is 100.}

\item{verbose}{Iterations at which the results are printed on the console. Default is every 100th iteration.}

\item{thin}{Thinning for gibbs updates. Default is 1 (no thinning).}

\item{reduce.mcmc}{The number of reduced MCMC iterations for marginal likelihood computations.
If \code{reduce.mcmc = NULL}, \code{mcmc/thin} is used.}

\item{ols.weight}{If TRUE, OLS estimates are used for adpative lasso weight vector.}

\item{c0}{Scale parameter for Gamma distribution. Used for the prior of \eqn{\sigma^2}.
Default is 0.1.}

\item{d0}{Shape parameter for Gamma distribution. Used for the prior of \eqn{\sigma^2}.
Default is 0.1.}

\item{nu.shape}{Shape parameter for Gamma distribution. Used for the prior for \eqn{\tau}.
Default is 2.0.}

\item{nu.rate}{Rate parameter for Gamma distribution. Used for the prior for \eqn{\tau}.
Default is 2.0.}

\item{known.alpha}{If \code{TRUE}, a user must specify a numeric value \eqn{[0, 2]} in \code{alpha.start}.
Default is \code{FALSE} and therefore \eqn{\alpha} will be estimated.}

\item{alpha.start}{Starting value for alpha.
When \code{known.alpha = TRUE}, alpha is fixed to the value of this argument.
Default is 1.}

\item{alpha.limit}{If \code{TRUE}, alpha is sampled from \eqn{(0,1)}, otherwise alpha is sampled between \eqn{(0,2)}.
Default is \code{FALSE}.}

\item{alpha.MH}{If \code{TRUE}, alpha is updated by the Metropolis–Hastings algorithm.
If \code{FALSE} the Griddy gibbs sampler is used instead.
Default is \code{TRUE}.}

\item{beta.start}{Starting values of beta. If \code{NULL}, randomly choose beta.start from OLS or standard normal distribution.
Default is \code{NULL}.}

\item{beta.alg}{An algorithm to sample beta.
Default is \code{beta.alg = "SVD"}.
Also supported is \code{beta.alg = "BCK"} and \code{beta.alg = "CHL"}.}

\item{regime.duration.min}{The minimum length of time in each regime. Default is 5.
If regime is shorter than this limit, hybrid analysis is skipped.}

\item{waic}{If \code{TRUE}, WAIC is computed after the parameter estimation.
Default is \code{FALSE}.}

\item{marginal}{If \code{TRUE}, the marginal likelihood is computed based on Chib's method.
Default is \code{FALSE}.}
}
\value{

}
\description{
Hybrid Univariate response linear change-point model with Bridge prior.
}
