% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeFixedPanel.R
\name{BridgeFixedPanel}
\alias{BridgeFixedPanel}
\title{Sparse Change Point Model with Fixed Effect}
\usage{
BridgeFixedPanel(formula, data, index, model, effect, n.break = 1,
  mcmc = 100, burn = 100, verbose = 100, thin = 1, b0 = 0, B0 = 1,
  c0 = 0.1, d0 = 0.1, r0 = 1, R0 = 1, nu.shape = 2, nu.rate = 2,
  alpha = 1, Waic = FALSE, marginal = FALSE)
}
\arguments{
\item{data}{Data.frame object.}

\item{index}{String vector for unit and time index variables.
For example, \code{index = c("unit", "year")}.}

\item{model}{Model (\code{c("within","between", "pooling")}).}

\item{effect}{Effect (\code{c("individual", "time", "twoways")}).}

\item{n.break}{Number of breaks.
If \code{n.break = 0}, it simply runs fixed effect model with shrinkage prior on coefficients.}

\item{mcmc}{MCMC iteration.}

\item{burn}{Burn-in period.}

\item{verbose}{Verbose.}

\item{thin}{Thinning.}

\item{c0}{Hyperparam}

\item{d0}{= 0.1}

\item{nu.shape}{=2.0}

\item{nu.rate}{=2.0}

\item{alpha}{= 1}

\item{fomula}{Inherited from \code{lm}. For example, \code{Y ~ X + Z}.}
}
\description{
Sparse Change Point Model with Fixed Effect
}
\author{
Jong Hee Park, and Soichiro Yamauchi \email{syamauchi@princeton.edu}
}
