% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeChangeMixedPanel.R
\name{BridgeMixedPanel}
\alias{BridgeMixedPanel}
\title{Mixed Effect Panel Model with Change-point}
\usage{
BridgeMixedPanel(
  y,
  X,
  W,
  subject.id,
  time.id,
  standardize = TRUE,
  n.break = 1,
  intercept = FALSE,
  mcmc = 100,
  burn = 100,
  verbose = 100,
  thin = 1,
  b0,
  B0,
  c0 = 0.1,
  d0 = 0.1,
  r0,
  R0,
  a = NULL,
  b = NULL,
  nu.shape = 2,
  nu.rate = 2,
  alpha = 1,
  alpha.MH = FALSE,
  beta.start = NULL,
  sigma2.start = NA,
  D.start = NA,
  P.start = NA,
  Waic = FALSE,
  marginal = FALSE,
  fixed = TRUE,
  beta.alg = "CHL",
  unscaled.Y,
  unscaled.X
)
}
\arguments{
\item{y}{Reponse vector}

\item{X}{Design matrix for fixed effects}

\item{W}{Design matrix for random effects.}

\item{subject.id}{unit id.}

\item{time.id}{time id.}

\item{standardize}{Default is \code{TRUE}.
If \code{TRUE}, Y will be demeaned, and X and W will be scaled to be mean zero and variance one.
Estimated parameters are always in the transformed scale.}

\item{n.break}{Number of structural breaks (changepoints). If \code{n.break = 0} is specified,
usual panel model is run with shrinkage prior on coefficients.}

\item{intercept}{boolean; if \code{FALSE}, the intercept is set to zero.}

\item{mcmc}{=100}

\item{burn}{=100}

\item{verbose}{=100 Verbose}

\item{thin}{Thinning}

\item{b0}{Hyperparam}

\item{B0}{Hyperparam}

\item{c0}{= 0.1}

\item{d0}{= 0.1}

\item{r0}{Hyperparam}

\item{R0}{Hyperparam}

\item{a}{= NULL}

\item{b}{= NULL}

\item{nu.shape}{=2.0}

\item{nu.rate}{=2.0}

\item{alpha}{= 1}

\item{alpha.MH}{If \code{TRUE}, alpha is updated by MH algorithm. By default, it is updated by griddy gibbs.}

\item{beta.start}{\code{= NA}}

\item{sigma2.start}{= NA}

\item{P.start}{= NA}

\item{Waic}{\code{= TRUE} If \code{= TRUE}, compute WAIC and store as "Waic.out." Retrieve using attr().}

\item{marginal}{\code{= FALSE}}

\item{fixed}{\code{= TRUE} If \code{= TRUE}, the fixed effects model is chosen as a baseline model.}

\item{beta.alg}{An algorithm to sample beta.
Default is \code{beta.alg = "BCK"}.
Also supported is the traditional sampler based on the Cholesky decomposition: \code{beta.alg = "CHL"}.}

\item{D.start=}{NA}
}
\value{
output
}
\description{
Linear mixed effect model with change-points.
}
