% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.BridgeChange}
\alias{coef.BridgeChange}
\title{Obtain posterior mean of coefficients}
\usage{
\method{coef}{BridgeChange}(out)
}
\arguments{
\item{out}{\code{BridgeChange} objects. Typically output from \code{BridgeChangeReg}.}
}
\description{
Obtain posterior mean of coefficients
}
