% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotRegime.R
\name{dotplotRegime}
\alias{dotplotRegime}
\title{Draw a plot of covariate-specific time-varying movements}
\usage{
dotplotRegime(
  out,
  hybrid = TRUE,
  start = 1,
  cex = 1,
  smooth.beta = TRUE,
  x.location = c("random", "legend", "default"),
  order.state = 1,
  location.bar = 9,
  text.cex = 1,
  legend.position = "topright",
  pos = 1,
  select = NULL,
  main = "",
  raw = FALSE
)
}
\arguments{
\item{out}{output of HMBB}

\item{hybrid}{If TRUE, DSS results are used for graph. Otherwise, HMBB results are used.}

\item{start}{Starting value of the time sequence}

\item{smooth.beta}{If TRUE, probabilistic estimates of hidden states (smooth) are used. Otherwise, discrete state values are used.}

\item{x.location}{x location of covariate name labels. If random, a random value (between min and max of x) is selected.
If legend, it will be displayed in the top-right as a legend. If default, the labels will be shown in the above of the last regime estimates.}

\item{thin}{The thinning interval used in the simulation.  The number of
MCMC iterations must be divisible by this value.}

\item{verbose}{A switch which determines whether or not the progress of the
sampler is printed to the screen.  If \code{verbose} is greater than 0 the
iteration number, the \eqn{\beta} vector, and the error variance are
printed to the screen every \code{verbose}th iteration.}

\item{degree.normal}{A null model for degree correction. Users can choose "NULL", "eigen" or "Lsym."
"NULL" is no degree correction. "eigen" is a principal eigen-matrix consisting of
the first eigenvalue and the corresponding eigenvector. "
Lsym" is a modularity matrix. Default is "eigen."}

\item{UL.Normal}{Transformation of sampled U. Users can choose "NULL", "Normal" or "Orthonormal."
"NULL" is no normalization. "Normal" is the standard normalization.
"Orthonormal" is the Gram-Schmidt orthgonalization. Default is "NULL."}

\item{v0}{\eqn{v_0/2} is the shape parameter for the inverse
Gamma prior on variance parameters for V.
If \code{v0 = NULL}, a value is computed from a test run of \code{NetworkStatic}.}

\item{v1}{\eqn{v_1/2} is the scale parameter for the
inverse Gamma prior on variance parameters for V.
If \code{v1 = NULL}, a value is computed from a test run of \code{NetworkStatic}.}
}
\description{
Draw a plot of covariate-specific time-varying movements
}
