% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{plotWaic}
\alias{plotWaic}
\title{Draw a plot of WAIC}
\usage{
plotWaic(out, xlab = "Model", ylab = "WAIC", main = "WAIC Comparison")
}
\arguments{
\item{waic}{WaicCompare output}
}
\value{
R base plot
}
\description{
Draw a plot of WAIC
}
