% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{WaicCompare}
\alias{WaicCompare}
\title{Compare Models based on WAIC}
\usage{
WaicCompare(outlist, print = TRUE)
}
\arguments{
\item{print}{boolean; if \code{TRUE} selected model is printed.}

\item{output}{list of \code{\link{BridgeChange}} objects, whch are typically outputs from \code{BridgeChangeReg}.}
}
\value{
A vector of WAIC
}
\description{
Compare Models based on WAIC
}
